Move /find-wallet to /wallets/find-wallet
